Move LongListUtil to util package, make generic method instead of class
