+++ src/main/java/by/kiselevich/periodicals/command/user/ShowUserSubscriptionsCommand.java
@@ -4,12 +4,12 @@ import by.kiselevich.periodicals.command.Attribute;
 import by.kiselevich.periodicals.command.Command;
 import by.kiselevich.periodicals.command.Page;
 import by.kiselevich.periodicals.command.admin.DashboardPageOptionCommand;
-import by.kiselevich.periodicals.command.admin.LongListUtil;
 import by.kiselevich.periodicals.entity.Subscription;
 import by.kiselevich.periodicals.exception.ServiceException;
 import by.kiselevich.periodicals.factory.ServiceFactory;
 import by.kiselevich.periodicals.factory.SubscriptionToItsStatusMapFactory;
 import by.kiselevich.periodicals.service.subscription.SubscriptionService;
+import by.kiselevich.periodicals.util.LongListUtil;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -23,11 +23,9 @@ import static by.kiselevich.periodicals.util.HttpUtil.getLocalizedMessageFromRes
 public class ShowUserSubscriptionsCommand implements Command {
 
     private final SubscriptionService subscriptionService;
-    private final LongListUtil<Subscription> longListUtil;
 
     public ShowUserSubscriptionsCommand() {
         subscriptionService = ServiceFactory.getInstance().getSubscriptionService();
-        longListUtil = new LongListUtil<>();
     }
 
     @Override
@@ -36,7 +34,7 @@ public class ShowUserSubscriptionsCommand implements Command {
         try {
             String login = (String) req.getSession().getAttribute(Attribute.LOGIN.getValue());
             List<Subscription> subscriptionList = subscriptionService.getAllSubscriptionsByUserLogin(login);
-            subscriptionList = longListUtil.getSubListByPageFromRequest(req, subscriptionList);
+            subscriptionList = LongListUtil.getSubListByPageFromRequest(req, subscriptionList);
             Map<Subscription, String> subscriptionAndStatusMap = SubscriptionToItsStatusMapFactory.getSubscriptionAndStatusMap(subscriptionList);
             Map<Subscription, String> subscriptionAndStatusSortedMap = new TreeMap<>(Comparator.comparing(Subscription::getSubscriptionStartDate).reversed());
             subscriptionAndStatusSortedMap.putAll(subscriptionAndStatusMap);
