+++ src/main/java/by/kiselevich/periodicals/util/LongListUtil.java
@@ -1,4 +1,4 @@
-package by.kiselevich.periodicals.command.admin;
+package by.kiselevich.periodicals.util;
 
 import by.kiselevich.periodicals.command.Attribute;
 import by.kiselevich.periodicals.command.JspParameter;
@@ -8,20 +8,24 @@ import java.util.List;
 
 /**
  * Class for {@link by.kiselevich.periodicals.command.Command} implementations, to work with long list by pages
- * @param <T> entity to work with parametrized {@link List}
  */
-public class LongListUtil<T> {
+public class LongListUtil {
     public static final int DEFAULT_RECORD_COUNT_PER_PAGE = 50;
     public static final int DEFAULT_PAGE_NUMBER = 1;
+
+    private LongListUtil() {
+
+    }
 
     /**
      * Returns sublist from received {@code list} by page number, got from {@link HttpServletRequest}. <br>
      * If page invalid, return first page.
      * @param req {@link HttpServletRequest} for page number getting
      * @param list {@link List} to get sublist from
+     * @param <T> entity to work with parametrized {@link List}
      * @return {@link List}
      */
-    public List<T> getSubListByPageFromRequest(HttpServletRequest req, List<T> list) {
+    public static <T> List<T> getSubListByPageFromRequest(HttpServletRequest req, List<T> list) {
         int subscriptionCount = list.size();
         int pages = subscriptionCount / DEFAULT_RECORD_COUNT_PER_PAGE + (subscriptionCount % DEFAULT_RECORD_COUNT_PER_PAGE == 0 ? 0 : 1);
         int pageNumber = getPageNumberFromRequest(req);
@@ -39,7 +43,7 @@ public class LongListUtil<T> {
         return list;
     }
 
-    private int getPageNumberFromRequest(HttpServletRequest req) {
+    private static int getPageNumberFromRequest(HttpServletRequest req) {
         String pageNumberString = req.getParameter(JspParameter.PAGE.getValue());
         try {
             return Integer.parseInt(pageNumberString);
